refactor(ingest): extract PredictScoringService from IngestPredictionService (max-params 2/3)
IngestPredictionService injected 4 deps (surreal, embedder, predicateRegistry,
config). Split the conflict-scoring engine — conflict weights/thresholds
(built from config), candidate embedding, and per-semantics scoring math
(single_active / bitemporal) — into PredictScoringService (config, embedder).
IngestPredictionService keeps the DB/registry orchestration (surreal,
predicateRegistry, scoring), 3 deps, reading scoring.conflict for the
threshold gates. Shared types + pure helpers moved to predictor-internals.ts
to avoid a cycle. Public predict() contract unchanged.

Part of the max-params=3 god-class split program.
